Rename IndexRoute to BaseRoute
(Better naming)
